fix a avx flag not defined bug

+++ cpp/include/znn/layer/deconv_as_conv.hpp
@@ -150,7 +150,7 @@ public:
    void forward(float const* __restrict i, float* __restrict o, 
      float const* __restrict runtime_kernel, float const* __restrict runtime_bias, float const* __restrict additive_scale)
    {
-      for (int c = 0; c < convs.size(); c++) {
+      for (size_t c = 0; c < convs.size(); c++) {
          convs[c]->forward(i, o, kernels[c]->data(), runtime_bias, additive_scale); 
       }
    }
